SW-83 #time 2h fix authority

+++ src/main/kotlin/com/stafsus/api/dto/InvitationDto.kt
@@ -16,5 +16,5 @@ data class InvitationDto(
 	@field:ValueOfEnum(enumClass = Authority::class)
 	val authority: String? = null,
 	@NotEmpty
-	val channels: List<Int>? = null,
+	val channels: List<Long>? = null,
 )
